requirements.txt
langchain-astradb
langchain 
langchain-google-genai
google-generativeai
langchain-openai
datasets
pypdf
python-dotenv
flask

-e .